Conflicts:
	app/View/Diagnostics/index.ctp
	app/View/Elements/left_sidebar.ctp

+++ app/Controller/ChargesController.php
@@ -23,7 +23,7 @@ class ChargesController extends AppController {
  */
 	public function index() {
 		$this->Charge->recursive = 0;
-		$this->set('charges', $this->Paginator->paginate());
+		$this->set('charges', $this->paginate());
 	}
 
 /**
@@ -50,10 +50,10 @@ class ChargesController extends AppController {
 		if ($this->request->is('post')) {
 			$this->Charge->create();
 			if ($this->Charge->save($this->request->data)) {
-				$this->Session->setFlash(__('The charge has been saved.'));
-				return $this->redirect(array('action' => 'index'));
+				$this->Session->setFlash(__('The charge has been saved'), 'flash/success');
+				$this->redirect(array('action' => 'index'));
 			} else {
-				$this->Session->setFlash(__('The charge could not be saved. Please, try again.'));
+				$this->Session->setFlash(__('The charge could not be saved. Please, try again.'), 'flash/error');
 			}
 		}
 		$people = $this->Charge->Person->find('list');
@@ -70,15 +70,16 @@ class ChargesController extends AppController {
  * @return void
  */
 	public function edit($id = null) {
+        $this->Charge->id = $id;
 		if (!$this->Charge->exists($id)) {
 			throw new NotFoundException(__('Invalid charge'));
 		}
-		if ($this->request->is(array('post', 'put'))) {
+		if ($this->request->is('post') || $this->request->is('put')) {
 			if ($this->Charge->save($this->request->data)) {
-				$this->Session->setFlash(__('The charge has been saved.'));
-				return $this->redirect(array('action' => 'index'));
+				$this->Session->setFlash(__('The charge has been saved'), 'flash/success');
+				$this->redirect(array('action' => 'index'));
 			} else {
-				$this->Session->setFlash(__('The charge could not be saved. Please, try again.'));
+				$this->Session->setFlash(__('The charge could not be saved. Please, try again.'), 'flash/error');
 			}
 		} else {
 			$options = array('conditions' => array('Charge.' . $this->Charge->primaryKey => $id));
@@ -94,20 +95,23 @@ class ChargesController extends AppController {
  * delete method
  *
  * @throws NotFoundException
+ * @throws MethodNotAllowedException
  * @param string $id
  * @return void
  */
 	public function delete($id = null) {
+		if (!$this->request->is('post')) {
+			throw new MethodNotAllowedException();
+		}
 		$this->Charge->id = $id;
 		if (!$this->Charge->exists()) {
 			throw new NotFoundException(__('Invalid charge'));
 		}
-		$this->request->allowMethod('post', 'delete');
 		if ($this->Charge->delete()) {
-			$this->Session->setFlash(__('The charge has been deleted.'));
-		} else {
-			$this->Session->setFlash(__('The charge could not be deleted. Please, try again.'));
+			$this->Session->setFlash(__('Charge deleted'), 'flash/success');
+			$this->redirect(array('action' => 'index'));
 		}
-		return $this->redirect(array('action' => 'index'));
+		$this->Session->setFlash(__('Charge was not deleted'), 'flash/error');
+		$this->redirect(array('action' => 'index'));
 	}
 }
